# Conflicts:
#	client/src/main/java/il/cshaifasweng/OCSFMediatorExample/client/SessionManager.java

+++ client/src/main/java/il/cshaifasweng/OCSFMediatorExample/client/SessionManager.java
@@ -1,5 +1,12 @@
 package il.cshaifasweng.OCSFMediatorExample.client;
 
+import il.cshaifasweng.OCSFMediatorExample.entities.Customer;
+import il.cshaifasweng.OCSFMediatorExample.entities.Employee;
+import il.cshaifasweng.OCSFMediatorExample.entities.Product;
+import il.cshaifasweng.OCSFMediatorExample.entities.User;
+import il.cshaifasweng.OCSFMediatorExample.entities.Order;
+
+import java.math.BigDecimal;
 import il.cshaifasweng.OCSFMediatorExample.entities.*;
 
 import java.util.HashMap;
@@ -13,7 +20,7 @@ public class SessionManager {
     private List<Product> cart;
     private List<Order> orders = new ArrayList<>();
     private Order selectedOrder;
-    private double orderTotal;
+    private BigDecimal orderTotal;
 
     private SessionManager() {
         cart = new ArrayList<>();  // Initialize the cart
@@ -75,10 +82,10 @@ public class SessionManager {
     public Order getSelectedOrder() {
         return selectedOrder;
     }
-    public void setOrderTotal(double orderTotal) {
+    public void setOrderTotal(BigDecimal orderTotal) {
         this.orderTotal = orderTotal;
     }
-    public double getOrderTotal() {
+    public BigDecimal getOrderTotal() {
         return orderTotal;
     }
